requirements.txt
pytest==7.4.0
requests==2.32.4
allure-pytest==2.15.3